GetAttrName, GetPartitionKey, TableName are public
appittome/vendor/store needed a mechanism to Marshall just the HashKey.

To get an object by hash, the key needs to be specified.  A method that
takes an interface and the Entity type, can check to make sure the
interface is of the appropirate type, marshall the key into the
appropriate dynamodb.GetItemInput, and sumbit it to Get to load the
entire entity.

Perhaps this should eventually be part of dynaGo, but for now it's
stashed in vendor/store, and these methods need to be public.

ADDED Encode can now handle 1 layer of indirection
CHNGD getAttrName => GetAttrName
CHNGD getPartitionKey => GetPartitionKey
CHNGD tableName => TableName

Signed-off-by: Evan J Brunner <ej3@appitto.me>