CreateKeyMaker accepts pointers to struct
There are common situations where a method will accept an
interface, and not knowing whether that interface points to
a pointer to a struct, or a struct directly will want to
create a KeyMaker.

Rather than make someone using the lib decipher a panic on
"reflect.NumField called on non-struct" - the CreateKeyMaker
has simply been adapted to quietly accept either a pointer to
a type, or the type itself.

CHNGD: CreateKeyMaker accepts reflect.Ptr to struct type

Signed-off-by: Evan J Brunner <ej3@appitto.me>